Consolidate more methods in Frame (#9305)
This PR standardizes some more methods that were previously duplicated between different Frame classes into Frame. Many of the changes should be straightforward to review, with the main possible exception being the changes to serialization logic. Note that because the format of the serialized data has been standardized across different Frame types, objects serialized with old versions of cudf will be incompatible with the new format. I've included some backwards compatibility layers to be removed in a future release. Additionally, this PR adds some methods to Frame that are really only applicable to DataFrame and Series. Since some such methods already exist in Frame, I think it's fine to continue adding to those pending the eventual creation of a common IndexedFrame class.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)

Approvers:
  - Ashwin Srinath (https://github.com/shwina)
  - GALI PREM SAGAR (https://github.com/galipremsagar)

URL: https://github.com/rapidsai/cudf/pull/9305
